Lots were added

+++ src/main/java/com/main/sellplatform/controller/rest/UserRestController.java
@@ -7,14 +7,11 @@ import org.springframework.http.MediaType;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
-import java.sql.SQLException;
 import java.util.List;
 
 @RestController
 @RequestMapping("api")
 public class UserRestController {
-
-
     private final UserService userService;
 
     @Autowired
@@ -22,14 +19,11 @@ public class UserRestController {
         this.userService = userService;
     }
 
-
     @PreAuthorize("hasAnyAuthority('user:read')")
     @GetMapping(value = "/getUsers", produces = {MediaType.APPLICATION_JSON_VALUE})
     public List<User> getUsersJdbc(){
         return userService.getUsers();
-
     }
-
 
     @PreAuthorize("hasAnyAuthority('user:write')")
     @PostMapping(value = "/postUser", produces = {MediaType.APPLICATION_JSON_VALUE})
@@ -38,17 +32,15 @@ public class UserRestController {
         userService.saveUser(user);
     }
 
-
     @PreAuthorize("hasAnyAuthority('user:read')")
     @GetMapping(value = "/getUser{id}", produces = {MediaType.APPLICATION_JSON_VALUE})
-    public User getUser(@PathVariable Integer id){
+    public User getUser(@PathVariable Long id){
         return userService.getUser(id);
     }
 
     @PreAuthorize("hasAnyAuthority('user:delete')")
     @DeleteMapping(value = "/delete{id}", produces = {MediaType.APPLICATION_JSON_VALUE} )
-    public void deleteUser(@PathVariable Integer id){
+    public void deleteUser(@PathVariable Long id){
         userService.deleteUser(id);
     }
-
 }
